ai-attendance-system/backend/requirements.txt
# Core dependencies first
setuptools>=65.0.0
wheel>=0.38.0

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Database
sqlalchemy>=2.0.0
aiofiles>=23.2.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
email-validator>=2.1.0

# Data processing
pandas>=2.0.0
openpyxl>=3.1.0
numpy>=1.24.0,<2.0.0

# Image processing and face recognition
Pillow>=10.0.0
opencv-python>=4.8.0

# Face recognition (may need manual installation)
# face-recognition>=1.3.0
# dlib>=19.24.0

# Environment
python-dotenv>=1.0.0